requirements.txt
discord.py==2.0.1
requests==2.25.1
